Collects all linear algebra errors into linalg/exceptions.jl
