+++ base/linalg/exceptions.jl
@@ -1,0 +1,22 @@
+export LAPACKException, SingularException, PosDefException,
+    RankDeficientException, DimensionMismatch
+
+type LAPACKException <: Exception
+    info::BlasInt
+end
+
+type SingularException <: Exception
+    info::BlasInt
+end
+
+type PosDefException <: Exception
+    info::BlasInt
+end
+
+type RankDeficientException <: Exception
+    info::BlasInt
+end
+
+type DimensionMismatch <: Exception
+    name::ASCIIString
+end
